move impl nft errors from lp_coins into nft mod
