+++ mm2src/coins/nft/nft_errors.rs
@@ -1,6 +1,6 @@
 use crate::eth::GetEthAddressError;
 use crate::nft::storage::{CreateNftStorageError, NftStorageError};
-use crate::GetMyAddressError;
+use crate::{GetMyAddressError, WithdrawError};
 use common::HttpStatusCode;
 use derive_more::Display;
 use enum_from::EnumFromStringify;
@@ -33,6 +33,10 @@ pub enum GetNftInfoError {
     },
     #[display(fmt = "DB error {}", _0)]
     DbError(String),
+}
+
+impl From<GetNftInfoError> for WithdrawError {
+    fn from(e: GetNftInfoError) -> Self { WithdrawError::GetNftInfoError(e) }
 }
 
 impl From<SlurpError> for GetNftInfoError {
